Move component-level derived data to a selector in the component
