+++ src/content/containers/ProfileFilterNavigator.js
@@ -1,7 +1,19 @@
+import { createSelector } from 'reselect';
 import { connect } from 'react-redux';
 import * as actions from '../actions';
-import { getRangeFilterLabels } from '../reducers/url-state';
+import { getRangeFilters } from '../reducers/url-state';
 import FilterNavigatorBar from '../components/FilterNavigatorBar';
+
+const getRangeFilterLabels = createSelector(
+  getRangeFilters,
+  rangeFilters => {
+    const labels = rangeFilters.map(range => {
+      return `Range: ${(range.start / 1000).toFixed(2)}s–${(range.end / 1000).toFixed(2)}s`;
+    });
+    labels.unshift('Complete Profile');
+    return labels;
+  }
+);
 
 export default connect(state => {
   const items = getRangeFilterLabels(state);
